Pull parametric logic functions into their own module
