+++ wasm/logic/control.py
@@ -267,20 +267,3 @@ def call_indirect_op(config: Configuration) -> None:
         raise Trap("Function type mismatch.  Expected {function_type}.  Got {function.type}")
 
     _setup_call(config, function_address)
-
-
-def drop_op(config: Configuration) -> None:
-    logger.debug("%s()", config.instructions.current.opcode.text)
-
-    config.pop_operand()
-
-
-def select_op(config: Configuration) -> None:
-    logger.debug("%s()", config.instructions.current.opcode.text)
-
-    a, b, c = config.pop3_operands()
-
-    if a:
-        config.push_operand(c)
-    else:
-        config.push_operand(b)
